Thanks rust compiler :)
Honestly I kinda keyboard mashed until it worked, but it works! The
function `search` now handles both case-sensitive and case-insensitive
searches. My only remaining question in this area is speed --- we turn
every line into a String for consistent types, and I don't know if
that'll be a constant memory allocation or if it'll grow linearly with
the size of the input file. (I want it to be constant.)